Move useCurrentLanguage hook into its own file
