+++ h5p-bildetema/src/hooks/useSiteLanguage.ts
@@ -1,19 +1,5 @@
-import { useLocation } from "react-router-dom";
-import { attributeLanguages } from "../../../common/constants/languages";
-import { LanguageCode } from "../../../common/types/LanguageCode";
 import { useL10n } from "./useL10n";
 
 export const useSiteLanguage = (): string => {
   return useL10n("htmlLanguageCode");
-};
-
-export const useCurrentLanguage = (): string => {
-  const { pathname } = useLocation();
-
-  const currentLanguageCode: LanguageCode =
-    pathname.split("/").length >= 2
-      ? (pathname.split("/")[1] as LanguageCode)
-      : "nob";
-
-  return attributeLanguages[currentLanguageCode];
 };
